% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_vocabs.R
\name{check_ecoregion}
\alias{check_ecoregion}
\alias{check_stat_recs}
\alias{check_countries}
\title{Lists of Vocabularies}
\usage{
check_ecoregion(arg, stop.on.fail = TRUE)

check_stat_recs(arg, stop.on.fail = TRUE)

check_countries(arg, stop.on.fail = TRUE)
}
\arguments{
\item{arg}{one of country code, ICES statistical rectangle or ICES ecoregion}

\item{stop.on.fail}{logical, if TRUE function will call stop()}
}
\value{
logical
}
\description{
Get a list of vocabularies such as country codes or ecoregions.
}
\examples{
\donttest{
check_ecoregion("Celtic Seas")

check_stat_recs("40F1")

check_countries("DK")
check_countries(c("DK", "GB"))
}
}
