% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_benthis_parameters.R
\name{get_benthis_parameters}
\alias{get_benthis_parameters}
\title{Get gear characteristics for 'benthis' gear categories}
\usage{
get_benthis_parameters()
}
\value{
a data.frame of model parameters
}
\description{
Download a data.frame of gear contact models for
'benthis' gear categories used in calculating
the swept area of a fishing gear.
}
\examples{
\donttest{
benthis_pars <- get_benthis_parameters()
}

}
