% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_datsu_session.R
\name{get_datsu_session}
\alias{get_datsu_session}
\title{Download VMS and logbook data from a datsu session}
\usage{
get_datsu_session(sessionId)
}
\arguments{
\item{sessionId}{datsu session ID}
}
\value{
a list of data.frames of VMS and logbook data
}
\description{
RESTRICED.  Only core members of the ICES VMS datacall can acess this data.
Download a list of data.frames of VMS and logbook data from the ICES DATSU data checking facility.
}
