% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_datsu_session_detail.R
\name{get_datsu_session_detail}
\alias{get_datsu_session_detail}
\title{Download details for a DATSU session}
\usage{
get_datsu_session_detail(sessionId)
}
\arguments{
\item{sessionId}{DATSU session ID}
}
\value{
a list
}
\description{
RESTRICTED.  Only core members of the ICES VMS data call can access this data.
Download a list of information on a DATSU submission.
}
