% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_datsu_session_ve.R
\name{get_datsu_session_ve}
\alias{get_datsu_session_ve}
\title{Download VMS data from a DATSU session}
\usage{
get_datsu_session_ve(sessionId)
}
\arguments{
\item{sessionId}{DATSU session ID}
}
\value{
a data.frame of VMS data
}
\description{
RESTRICTED.  Only core members of the ICES VMS data call can access this data.
Download a data.frame of VMS data from the ICES DATSU data checking facility.
}
