% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_fo_effort.R
\name{get_fo_effort}
\alias{get_fo_effort}
\title{Download fishing effort summaries}
\usage{
get_fo_effort(ecoregion)
}
\arguments{
\item{ecoregion}{ICES ecoregion}
}
\value{
a data.frame
}
\description{
Download a data.frame of kw fishing hours by country and year for a
given ICES ecoregion.
}
\examples{
\dontrun{
# requires authorization
ns_effort_data <- get_fo_effort("Greater North Sea")
}

}
