% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_fo_landings.R
\name{get_fo_landings}
\alias{get_fo_landings}
\title{Download fishing landings summaries}
\usage{
get_fo_landings(ecoregion)
}
\arguments{
\item{ecoregion}{ICES ecoregion}
}
\value{
a data.frame
}
\description{
Download a data.frame of total weight by country and year for a
given ICES ecoregion.
}
\examples{
\dontrun{
# requires authorization
ns_landings_data <- get_fo_landings("Greater North Sea")
}

}
