% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_metier_lookup.R
\name{get_metier_lookup}
\alias{get_metier_lookup}
\title{Link Metier level 6 to Benthis categories}
\usage{
get_metier_lookup()
}
\value{
a data.frame
}
\description{
Download a data.frame of Metier codes to link
level 6 metier codes with different gear categories
}
\examples{
\donttest{
metier_lookup <- get_metier_lookup()
}

}
