% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_screening_details.R
\name{get_screening_details}
\alias{get_screening_details}
\title{Get the details of a file screening session}
\usage{
get_screening_details(sessionId)
}
\arguments{
\item{sessionId}{file screening session ID}
}
\value{
a list
}
\description{
Download a list of information on a file upload screening session.
}
\examples{
\donttest{
screening_info <- get_screening_details(530)
}
}
