% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_screening_messages.R
\name{get_screening_messages}
\alias{get_screening_messages}
\title{Get all messages of a file screening session}
\usage{
get_screening_messages(sessionId)
}
\arguments{
\item{sessionId}{file screening session ID}
}
\value{
a list
}
\description{
Download a list of error and warning messages on a file upload screening session.
}
\examples{
\donttest{
tofix <- get_screening_messages(528)
}
}
