% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_upload_summary.R
\name{get_upload_summary}
\alias{get_upload_summary}
\title{Download a summary of submitted data}
\usage{
get_upload_summary(datacall = NULL)
}
\arguments{
\item{datacall}{integer year giving which data call year to inquire about.
If NULL returns the a summary of the most recent approved data.}
}
\value{
a data.frame of VMS summary data
}
\description{
Download a summary of submitted data
}
\examples{
\dontrun{
upload_summary <- get_upload_summary()
}

}
