% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_vocab_lists.R
\name{get_ecoregion_list}
\alias{get_ecoregion_list}
\alias{get_stat_rec_list}
\alias{get_country_list}
\title{Lists of Vocabularies}
\usage{
get_ecoregion_list()

get_stat_rec_list()

get_country_list()
}
\value{
a character vector
}
\description{
Get a list of vocabularies such as country codes or ecoregions.
}
\examples{
\donttest{
ecoregions <- get_ecoregion_list()
"Celtic Seas" \%in\% ecoregions

stat_recs <- get_stat_rec_list()
"40F1" \%in\% stat_recs

countries <- get_country_list()
"DK" \%in\% countries
}
}
