% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/has_sf.R
\name{has_sf}
\alias{has_sf}
\title{Check to see if the sf package is installed}
\usage{
has_sf()
}
\value{
logical value, TRUE if the \code{sf} package is installed
}
\description{
check to see if the \code{sf} package is installed
}
\examples{

(has_sf())
}
