% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/icesVMS-package.R
\docType{package}
\name{icesVMS-package}
\alias{icesVMS-package}
\alias{icesVMS}
\title{Functions to link to the ICES VMS and logbook database web services.}
\description{
Functions to link to the ICES VMS and logbook database web services to allow
users to download summaries and data products from the ICES VMS and logbook database.
}
\details{
\emph{Download data:}
\tabular{ll}{
  \code{\link{get_benthis_parameters}}     \tab download gear contact model parameters\cr
  \code{\link{get_metier_lookup}} \tab download metier to gear category lookup table\cr
  \code{\link{get_vms}}         \tab download VMS data
}
}
\references{
ICES VMS and Logbook Database web services: \url{https://data.ices.dk/vms/webservices}.
}
\author{
Colin Millar.
}
