% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update_token.R
\name{update_token}
\alias{update_token}
\title{Update user token for project}
\usage{
update_token(user)
}
\arguments{
\item{user}{a character string giving the users ICES username}
}
\value{
TRUE or FALSE depenting on the success of the operation
}
\description{
Must be run in interactive mode.  User is promted for a password, and if successful
a Javascript Web Token (JWT) is saved to the project root (current working directory).
The token will last for 1 hour, after which the user needs to update the token using 
this function again.
}
