% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vms_api.R
\name{vms_api}
\alias{vms_api}
\title{Build a VMS web service url}
\usage{
vms_api(service, ...)
}
\arguments{
\item{service}{the name of the service}

\item{...}{named arguments will be added as queries}
}
\value{
a complete url as a character string
}
\description{
utility to build a url with optional query arguments
}
\examples{

vms_api("hi", bye = 21)
vms_api("csquares", argument1 = 2021, argument2 = NULL)

}
