% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wkt_csquare.R
\name{wkt_csquare}
\alias{wkt_csquare}
\title{Build a WKT string from latitude and longitude}
\usage{
wkt_csquare(lat, lon)
}
\arguments{
\item{lat}{latitude}

\item{lon}{longitude}
}
\value{
a string in WKT format
}
\description{
utility to build a WKT (Well Known Text) polygon string
for a c-square
}
\examples{

wkt_csquare(55, 0.1)

}
