% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_csquare.R
\name{get_csquare}
\alias{get_csquare}
\title{Download VMS data}
\usage{
get_csquare(c_square, stat_rec, ices_area, ecoregion)
}
\arguments{
\item{c_square}{character 0.05 degree c-square name}

\item{stat_rec}{ICES statistical rectangle}

\item{ices_area}{ICES area}

\item{ecoregion}{ICES ecoregion}

\item{country}{country code}

\item{year}{integer year}

\item{month}{integer month}

\item{gear_code}{benthis gear code}

\item{metier}{level 6 metier code}
}
\value{
a data.frame of VMS data
}
\description{
RESTRICED.  Only core members of the ICES VMS datacall can acess this data.
Download a data.frame of VMS data from the ICES VMS and logbook database.
}
