% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_datsu_session_le.R
\name{get_datsu_session_le}
\alias{get_datsu_session_le}
\title{Download logbook data from a DATSU session}
\usage{
get_datsu_session_le(sessionId)
}
\arguments{
\item{sessionId}{DATSU session ID}
}
\value{
a data.frame of logbook data
}
\description{
RESTRICTED.  Only core members of the ICES VMS data call can access this data.
Download a data.frame of logbook data from the ICES DATSU data checking facility.
}
