% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_effort_map.R
\name{get_effort_map}
\alias{get_effort_map}
\title{Download fishing effort map data}
\usage{
get_effort_map(ecoregion, year = NULL)
}
\arguments{
\item{ecoregion}{ICES ecoregion}

\item{year}{which year to select (see details)}
}
\value{
a data.frame with a WKT column for the c-square polygons
}
\description{
Download a data.frame of kw fishing hours by c-square and gear
category averaged over 4 years.
}
\details{
The spatial data.frame contains average annual mw fishing hours,
averaged over 4 years.
}
