% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_logbook.R
\name{get_logbook}
\alias{get_logbook}
\title{Download Logbook data}
\usage{
get_logbook(
  country,
  year,
  month,
  gear_code,
  metier,
  stat_rec,
  ices_area,
  ecoregion
)
}
\arguments{
\item{country}{country code}

\item{year}{integer year}

\item{month}{integer month}

\item{gear_code}{benthis gear code}

\item{metier}{level 6 metier code}

\item{stat_rec}{ICES statistical rectangle}

\item{ices_area}{ICES area}

\item{ecoregion}{ICES ecoregion}
}
\value{
a data.frame of VMS data
}
\description{
RESTRICED.  Only core members of the ICES VMS datacall can acess this data.
Download a data.frame of VMS data from the ICES VMS and logbook database.
}
