% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_passive_footprint.R
\name{get_passive_footprint}
\alias{get_passive_footprint}
\title{Download passive fishing gear footprint}
\usage{
get_passive_footprint(ecoregion, year, metier_level4 = NULL)
}
\arguments{
\item{ecoregion}{ICES ecoregion}

\item{year}{which year to select}

\item{metier_level4}{optional gear code (metier level 4) ("FPO")}
}
\value{
a data.frame with a WKT column for the c-square polygons
}
\description{
Download a data.frame of presence of fishing by c-square and year
for passive fishing gears (see details).
}
\details{
Passive gears defined as all gears registered under
the metier level 4 codes, FPO (fishing pots), LLS (long lines) and
GNS (set gill nets), with the exclusion of metier level 5 codes
within the GNS category: GNS_SPF and GNS_LPF (set gill nets
targeting small and large pelagic fish).
}
