% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_sar_map.R
\name{get_sar_map}
\alias{get_sar_map}
\title{Download fishing effort summaries}
\usage{
get_sar_map(ecoregion, year = NULL)
}
\arguments{
\item{ecoregion}{ICES ecoregion}

\item{year}{which year to select (see details)}
}
\value{
a data.frame with a WKT column for the c-square polygons
}
\description{
Download a data.frame of kw fishing hours by country and year for a
given ICES ecoregion.
}
\details{
The spatial data.frame contains average annual mw fishing hours,
averaged over 4 years.
}
