% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_sar_map.R
\name{get_sar_map}
\alias{get_sar_map}
\title{Download swept area ratio map data}
\usage{
get_sar_map(ecoregion, year = NULL)
}
\arguments{
\item{ecoregion}{ICES ecoregion}

\item{year}{which year to select (see details)}
}
\value{
a data.frame with a WKT column for the c-square polygons
}
\description{
Download a data.frame of surface and subsurface swept area 
ratio by c-square for a given ICES ecoregion.
}
\details{
The spatial data.frame contains average annual surface-sar 
and subsurface-sar averaged over 4 years.
}
