% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_wgfbit_data1.R
\name{get_wgfbit_data1}
\alias{get_wgfbit_data1}
\title{Download swept area ratio, landings and value map data}
\usage{
get_wgfbit_data1(ecoregion, year, fishing_category = NULL,
  benthis_metier = NULL)
}
\arguments{
\item{ecoregion}{ICES ecoregion}

\item{year}{which year to select}

\item{fishing_category}{optional gear category ("Otter", "Dredge")}

\item{benthis_metier}{optional benthis metier ("SDN_DMF")}
}
\value{
a data.frame with a WKT column for the c-square polygons
}
\description{
Download a data.frame of surface and subsurface swept area 
ratio by c-square for a given ICES ecoregion, year and gear code.
}
\details{
gear_group and benthis_metier may not both be supplied, if neither
are supplied the total is calculated.
}
