% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_wgfbit_data2.R
\name{get_wgfbit_data2}
\alias{get_wgfbit_data2}
\title{Download swept area ratio, landings and value map data}
\usage{
get_wgfbit_data2(ecoregion, year)
}
\arguments{
\item{ecoregion}{ICES ecoregion}

\item{year}{which year to select}
}
\value{
a data.frame with a WKT column for the c-square polygons
}
\description{
Download a data.frame of surface and subsurface swept area
ratio by c-square for a given ICES ecoregion, year and gear code.
}
