% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/icesVMS-package.R
\docType{package}
\name{icesVMS-package}
\alias{icesVMS-package}
\alias{icesVMS}
\title{Functions to link to the ICES VMS and logbook database webservices.}
\description{
Functions to link to the ICES VMS and logbook database webservices to allow
users to download summaries and data products from the ICES VMS and logbook database.
}
\details{
to come later
}
\references{
ICES web services: \url{https://www.ices.dk/marine-data/tools/Pages/WebServices.aspx}.
}
\author{
Colin Millar.
}
