% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/icesVMS-package.R
\docType{package}
\name{icesVMS-package}
\alias{icesVMS-package}
\alias{icesVMS}
\title{Functions to link to the ICES VMS and logbook database webservices.}
\description{
Functions to link to the ICES VMS and logbook database webservices to allow
users to download summaries and data products from the ICES VMS and logbook database.
}
\details{
\emph{Download data:}
\tabular{ll}{
  \code{\link{get_benthis_parameters}}     \tab download gear contact model parameters\cr
  \code{\link{get_metier_lookup}} \tab download metier to gear category lookup table\cr
  \code{\link{get_vms}}         \tab download VMS data
}
\emph{Authentication:}
\tabular{ll}{
  \code{\link{update_token}}     \tab update the token for the current project
}
}
\references{
ICES web services: \url{https://www.ices.dk/marine-data/tools/Pages/WebServices.aspx}.
}
\author{
Colin Millar.
}
