% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/screen_vms_file.R
\name{screen_vms_file}
\alias{screen_vms_file}
\title{Download swept area ratio, landings and value map data}
\usage{
screen_vms_file(file)
}
\arguments{
\item{file}{file name of the file containing the data to screen}
}
\value{
text message from the screening api
}
\description{
In this webservice the user can upload a file to be screen and
validated by the VMS database. The service can be called using post
and the file will have to be part of the body of que call. The user
needs to be autenticated in order to call this service.
This file can be later pushed to the database by the same user.
}
